+++ src/components/App.jsx
@@ -1,8 +1,7 @@
 import React from "react";
-import Filters from "./Filters/Filters";
-import MoviesList from "./Movies/MoviesList";
 import Header from "./Header/Header";
 import LoginModal from "./Modals/LoginModal";
+import MoviesPage from "./pages/MoviesPage/MoviesPage";
 import CallApi from "../api/api";
 import _ from "lodash";
 import Cookies from "universal-cookie";
@@ -10,13 +9,6 @@ import Cookies from "universal-cookie";
 const cookies = new Cookies();
 
 export const AppContext = React.createContext();
-
-
-const initialFilters = {
-  sort_by: "popularity.desc",
-  primary_release_year: "primary_release_year",
-  with_genres: [],
-};
 
 export default class App extends React.Component {
   constructor(props) {
@@ -27,9 +19,6 @@ export default class App extends React.Component {
       favorites: [],
       watchlist: [],
       session_id: null,
-      filters: initialFilters,
-      page: 1,
-      total_pages: 1,
       showLoginModal: false,
     }
   }
@@ -91,37 +80,6 @@ export default class App extends React.Component {
     })
   };
 
-  onChangeFilters = event => {
-    const value = event.target.value;
-    const name = event.target.name;
-
-    this.setState(prevState => ({
-      filters: {
-        ...prevState.filters,
-        [name]: value
-      }
-    }));
-  };
-
-  onChangePage = page => {
-    this.setState({
-      page
-    })
-  };
-
-  onChangeTotalPages = value => {
-    this.setState({
-      total_pages: value
-    })
-  };
-
-  onResetFilters = () => {
-    this.setState({
-      filters: initialFilters,
-      page: 1
-    })
-  };
-
   getWatchlist = () => {
     const {session_id, user} = this.state;
 
@@ -154,10 +112,6 @@ export default class App extends React.Component {
 
   render() {
     const {
-      filters,
-      page,
-      total_pages,
-      genres,
       user,
       session_id,
       favorites,
@@ -186,36 +140,7 @@ export default class App extends React.Component {
             user={user}
             updateSessionId={this.updateSessionId}
           />
-          <div className="container pt-1">
-            <div className="row">
-              <div className="col-4">
-                <div className="card sticky-top">
-                  <div className="card-body py-1">
-                    <h3>Фильтры:</h3>
-                    <Filters
-                      filters={filters}
-                      page={page}
-                      total_pages={total_pages}
-                      genres={genres}
-                      onChangeFilters={this.onChangeFilters}
-                      onChangePage={this.onChangePage}
-                      onResetFilters={this.onResetFilters}
-                    />
-                  </div>
-                </div>
-              </div>
-              <div className="col-8">
-                <MoviesList
-                  filters={filters}
-                  page={page}
-                  session_id={session_id}
-                  onChangePage={this.onChangePage}
-                  onChangeTotalPages={this.onChangeTotalPages}
-                  toggleModal={this.toggleModal}
-                />
-              </div>
-            </div>
-          </div>
+          <MoviesPage/>
           {
             showLoginModal && <LoginModal/>
           }
